+++ src/widgets/containers/app_root.rs
@@ -179,19 +179,19 @@ impl Widget for AppRoot {
             let style_constraints = constraints_from_style(&resolved);
             let constraints = merge_constraints(style_constraints, child.layout_constraints());
             let render_width = clamp_with_constraints(
-                width.saturating_sub(margin.left + margin.right).max(1),
+                width.saturating_sub(margin.left as usize + margin.right as usize).max(1),
                 constraints.min_width,
                 constraints.max_width,
-                width.saturating_sub(margin.left + margin.right).max(1),
+                width.saturating_sub(margin.left as usize + margin.right as usize).max(1),
             );
             let render_height = clamp_with_constraints(
                 height_limit
-                    .saturating_sub(margin.top + margin.bottom)
+                    .saturating_sub(margin.top as usize + margin.bottom as usize)
                     .max(1),
                 constraints.min_height,
                 constraints.max_height,
                 height_limit
-                    .saturating_sub(margin.top + margin.bottom)
+                    .saturating_sub(margin.top as usize + margin.bottom as usize)
                     .max(1),
             );
             let render_height = if let Some(fixed_total) = child.layout_height() {
@@ -267,19 +267,19 @@ impl Widget for AppRoot {
             let style_constraints = constraints_from_style(&resolved);
             let constraints = merge_constraints(style_constraints, child.layout_constraints());
             let render_width = clamp_with_constraints(
-                width.saturating_sub(margin.left + margin.right).max(1),
+                width.saturating_sub(margin.left as usize + margin.right as usize).max(1),
                 constraints.min_width,
                 constraints.max_width,
-                width.saturating_sub(margin.left + margin.right).max(1),
+                width.saturating_sub(margin.left as usize + margin.right as usize).max(1),
             );
             let render_height = clamp_with_constraints(
                 height_limit
-                    .saturating_sub(margin.top + margin.bottom)
+                    .saturating_sub(margin.top as usize + margin.bottom as usize)
                     .max(1),
                 constraints.min_height,
                 constraints.max_height,
                 height_limit
-                    .saturating_sub(margin.top + margin.bottom)
+                    .saturating_sub(margin.top as usize + margin.bottom as usize)
                     .max(1),
             );
             let render_height = if let Some(fixed_total) = child.layout_height() {
@@ -446,7 +446,7 @@ impl Widget for AppRoot {
                 Some(height) => {
                     total = total
                         .saturating_add(height)
-                        .saturating_add(margin.top + margin.bottom);
+                        .saturating_add(margin.top as usize + margin.bottom as usize);
                 }
                 None => return None,
             }
@@ -462,7 +462,7 @@ impl Widget for AppRoot {
             let resolved = css::resolve_style(child.as_ref(), &meta);
             let margin = margin_from_style(&resolved);
             if let Some(width) = child.content_width() {
-                widest = widest.max(width.saturating_add(margin.left + margin.right));
+                widest = widest.max(width.saturating_add(margin.left as usize + margin.right as usize));
                 any = true;
             }
         }
